backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
requests
openai
python-multipart
pytest
httpx
